+++ tests/dummy/__init__.py
@@ -1,4 +1,0 @@
-class DoNotLoadClass:
-    # This class should not be loaded by the PluginRegistry
-    def __init__(self):
-        pass