+++ tests/dummy/test.py
@@ -1,7 +1,0 @@
-class DummyClass:
-    def __init__(self, name: str = None):
-        self.name = name
-        pass
-
-
-dummy_instantiated = DummyClass(name='dummy')